test(steps): add tests
